Update thermostatic_fan.md
This edit is driven by the fact that the original RC config.g file disables Thermostatic fans in the first line to delete, and then sets that fan to 0 at start with the second. The edit allows total control using a single line of code, and is more clear to the end user on how to properly configure this customization.